+++ client/components/Footer/Footer.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import {Link} from 'react-router';
+
+class Footer extends React.Component {
+  render() {
+    return (
+      <footer>
+        <ul>
+          <li><Link to="/">Dashboard</Link></li>
+          <li><Link to="/profile">Profile</Link></li>
+          <li><Link to="/help">Help</Link></li>
+        </ul>
+      </footer>
+    );
+  }
+}
+
+export default Footer;